move type files under a types/ dir
https://coveord.atlassian.net/browse/KIT-2600
