+++ packages/headless/src/app/ssr-engine/types/build-ssr-types.ts
@@ -1,4 +1,4 @@
-import {CoreEngine} from '../engine';
+import {CoreEngine} from '../../engine';
 import {
   ControllersMap,
   ControllersPropsMap,
