+++ packages/headless/src/app/ssr-engine/types/fetch-initial-state-ssr-types.ts
